# fix: prettify CODE_GEN_OUTPUT_DIR to improve performance