--- a/scripts/generate-clients/code-prettify.js
+++ b/scripts/generate-clients/code-prettify.js
@@ -4,6 +4,7 @@ const { spawnProcess } = require("./spawn-process");
 const { CODE_GEN_OUTPUT_DIR } = require("./code-gen-dir");
 
 const prettifyCode = async () => {
+  const spawnPromiseArray = [];
   for (const modelName of readdirSync(CODE_GEN_OUTPUT_DIR)) {
     if (modelName === "source") continue;
     const artifactPath = path.join(
@@ -11,11 +12,14 @@ const prettifyCode = async () => {
       modelName,
       "typescript-codegen"
     );
-    await spawnProcess("./node_modules/.bin/prettier", [
-      "--write",
-      `${artifactPath}/**/*.{ts,md,json}`
-    ]);
+    spawnPromiseArray.push(
+      spawnProcess("./node_modules/.bin/prettier", [
+        "--write",
+        `${artifactPath}/**/*.{ts,md,json}`
+      ])
+    );
   }
+  await Promise.all(spawnPromiseArray);
 };
 
 module.exports = {