*Issue #, if available:*
Fixes https://github.com/aws/aws-sdk-js-v3/issues/707

*Description of changes:*
* prettifies CODE_GEN_OUTPUT_DIR instead of each package to improve performance
* Tested on `c5.4xlarge` EC2 instance for generating 99 aws.json-1.1 clients:
  * without this code change:
  ```console
  $ time yarn generate-clients -m <dir-containing-json-1.1-models>
  ...
  ...
  Done in 205.55s.
  284.17s user 13.39s system 144% cpu 3:25.72 total
  ```
  * with this code change:
  ```console
  $ time yarn generate-clients -m <dir-containing-json-1.1-models>
  ...
  ...
  Done in 132.67s.
  177.20s user 7.31s system 138% cpu 2:12.84 total
  ```

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.